requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
gunicorn==25.1.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
python-dotenv==1.0.1
